Fix: respect feat_smartpls setting in playlist_delete_all function

+++ src/mpd_client/mpd_client_playlists.c
@@ -336,30 +336,32 @@ sds mpd_client_playlist_delete_all(t_config *config, t_mpd_client_state *mpd_cli
         return buffer;
     }
     //delete each smart playlist file that have no corresponding mpd playlist file
-    sds smartpls_path = sdscatfmt(sdsempty(), "%s/smartpls", config->varlibdir);
-    DIR *smartpls_dir = opendir(smartpls_path);
-    if (smartpls_dir != NULL) {
-        struct dirent *next_file;
-        while ((next_file = readdir(smartpls_dir)) != NULL ) {
-            if (strncmp(next_file->d_name, ".", 1) != 0) {
-                if (list_get_node(&playlists, next_file->d_name) == NULL) {
-                    sds smartpls_file = sdscatfmt(sdsempty(), "%s/smartpls/%s", config->varlibdir, next_file->d_name);
-                    if (unlink(smartpls_file) == 0) {
-                        LOG_VERBOSE("Removed orphaned smartpls file \"%s\"", smartpls_file);
+    if (mpd_client_state->feat_smartpls == true) {
+        sds smartpls_path = sdscatfmt(sdsempty(), "%s/smartpls", config->varlibdir);
+        DIR *smartpls_dir = opendir(smartpls_path);
+        if (smartpls_dir != NULL) {
+            struct dirent *next_file;
+            while ((next_file = readdir(smartpls_dir)) != NULL ) {
+                if (strncmp(next_file->d_name, ".", 1) != 0) {
+                    if (list_get_node(&playlists, next_file->d_name) == NULL) {
+                        sds smartpls_file = sdscatfmt(sdsempty(), "%s/smartpls/%s", config->varlibdir, next_file->d_name);
+                        if (unlink(smartpls_file) == 0) {
+                            LOG_VERBOSE("Removed orphaned smartpls file \"%s\"", smartpls_file);
+                        }
+                        else {
+                            LOG_ERROR("Removing smartpls file \"%s\" failed: %s", smartpls_file, strerror(errno));
+                        }
+                        sdsfree(smartpls_file);
                     }
-                    else {
-                        LOG_ERROR("Removing smartpls file \"%s\" failed: %s", smartpls_file, strerror(errno));
-                    }
-                    sdsfree(smartpls_file);
                 }
             }
+            closedir(smartpls_dir);
         }
-        closedir(smartpls_dir);
+        else {
+            LOG_ERROR("Can not open smartpls dir \"%s\"", smartpls_path);
+        }
+        sdsfree(smartpls_path);
     }
-    else {
-        LOG_ERROR("Can not open smartpls dir \"%s\"", smartpls_path);
-    }
-    sdsfree(smartpls_path);
     
     if (strcmp(type, "deleteEmptyPlaylists") == 0) {
         struct list_node *current = playlists.head;
@@ -373,7 +375,7 @@ sds mpd_client_playlist_delete_all(t_config *config, t_mpd_client_state *mpd_cli
         struct list_node *current = playlists.head;
         while (current != NULL) {
             bool smartpls = false;
-            if (strcmp(type, "deleteSmartPlaylists") == 0) {
+            if (mpd_client_state->feat_smartpls == true && strcmp(type, "deleteSmartPlaylists") == 0) {
                 sds smartpls_file = sdscatfmt(sdsempty(), "%s/smartpls/%s", config->varlibdir, current->key);
                 if (unlink(smartpls_file) == 0) {
                     LOG_VERBOSE("Smartpls file %s removed", smartpls_file);
